Add files via upload
Description: Study Hours and Student Scores Prediction
This project involves the application of a simple linear regression model in predicting the score of students based on the number of hours put in studying. In this example, one will be looking at how to set up a relationship between study hours and student performance and predict student scores based on study hours.

Dataset
In this dataset, one is working with the following:

Hours: number of hours a student studied
Scores: the corresponding score or percentage achieved by the student
Steps Involved:
Data Preparation:

The dataset is given in dictionary format; therefore, it will be transformed into a Pandas DataFrame for easy handling and analysis.
Extract and prepare the independent variable, Hours, and dependent variable, Scores.
Data Splitting:

It means that the data has to be split 80:30 for the training and testing sets. In other words, the model will be trained on some subset of the data and then tested on other unseen data to see how it works.
Model Training:

A linear regression model from the sklearn library is created with class LinearRegression.
Using the training data (X_train, y_train), train the model to create a relationship between study hours and score.
Model Evaluation:

In this case, the model's predictions on the test set, X_test, are scored using the following important metrics: 
Mean Absolute Error: It is the average magnitude of the errors in predictions.
Mean Squared Error: It is the average of the squared differences between predicted and actual values.
R-squared : It explains how well the regression line fits the actual data points.
Visualization:

It graphs, by hours studied versus scores achieved, a scatter plot, with the regression line overlayed to show the trend.
Prediction Function:
 A function predict_score allows the user to provide an input of hours of studying and returns a predicted score based on the fitted linear regression model above.
Example Prediction:

It comes with an example usage in which it is forecasting the score for a student who studies 5 hours. One can change this around to forecast scores for any number of study hours.